requirements.txt
flask
requests==2.31.0
requests-cache==1.2.1
pandas==1.5.3
plotly==5.9.0
python-dateutil
lxml
html5lib
feedparser==6.0.12
beautifulsoup4